Move the outbound mail relay under mail as relay
There were two unrelated things called relay: the frp tunnel that carries app
traffic, and the smtp relay that hands outbound mail to ses. Talking about
either meant saying which one every time.

Nesting settles it by position rather than by a longer name. The package is
relay inside mail, next to the listener helpers it already shares. The two
files that need both import this one under an alias, which is what the
ambiguity actually costs.

Pure move, no behaviour change.
