+++ backend/cmd/api/main.go
@@ -8,8 +8,8 @@ import (
 	"github.com/syncloud/redirect/dns"
 	"github.com/syncloud/redirect/ioc"
 	"github.com/syncloud/redirect/log"
+	mailrelay "github.com/syncloud/redirect/mail/relay"
 	"github.com/syncloud/redirect/mailin"
-	"github.com/syncloud/redirect/mailrelay"
 	"github.com/syncloud/redirect/metrics"
 	"github.com/syncloud/redirect/relay"
 	"github.com/syncloud/redirect/rest"
