+++ backend/ioc/ioc.go
@@ -16,8 +16,8 @@ import (
 	"github.com/syncloud/redirect/dns"
 	"github.com/syncloud/redirect/log"
 	"github.com/syncloud/redirect/mail"
+	mailrelay "github.com/syncloud/redirect/mail/relay"
 	"github.com/syncloud/redirect/mailin"
-	"github.com/syncloud/redirect/mailrelay"
 	"github.com/syncloud/redirect/metrics"
 	"github.com/syncloud/redirect/probe"
 	"github.com/syncloud/redirect/relay"
