+++ backend/mail/relay/account_usage.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import "time"
