+++ backend/mail/relay/cloud_watch.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"time"
