+++ backend/mail/relay/db_store.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import "time"
