+++ backend/mail/relay/feedback.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"encoding/json"
