+++ backend/mail/relay/relay_test.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"net/http"
