+++ backend/mail/relay/reputation.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"sync"
