+++ backend/mail/relay/reputation_test.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"fmt"
