+++ backend/mail/relay/rspamd.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"bytes"
