+++ backend/mail/relay/rspamd_result.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 type rspamdResult struct {
 	Action string  `json:"action"`
