+++ backend/mail/relay/ses_sender.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
