+++ backend/mail/relay/ses_test.go
@@ -1,4 +1,4 @@
-package mailrelay
+package relay
 
 import (
 	"bytes"
